Add an option to specify the Highlight.js instance and provide a core plugin
