+++ README.md
@@ -1,4 +1,4 @@
-# markdown-it-highlightjs [![npm version](http://img.shields.io/npm/v/markdown-it-highlightjs.svg?style=flat-square)](https://www.npmjs.org/package/markdown-it-highlightjs)
+# markdown-it-highlightjs [![npm version](https://img.shields.io/npm/v/markdown-it-highlightjs.svg?style=flat-square)](https://www.npmjs.org/package/markdown-it-highlightjs)
 
 > Preset to use [highlight.js] with [markdown-it].
 
@@ -23,6 +23,7 @@ Name       | Type | Description                                                 
 `code`     | boolean | Whether to add the `hljs` class to raw code blocks (not fenced blocks). | `true`
 `register` | object  | Register other languages which are not included in the standard pack.   | `null`
 `inline`   | boolean | Whether to highlight inline code.                                       | `false`
+`hljs`     | object  | Provide the instance of [highlight.js] to use for highlighting          | `require('highlight.js')`
 
 ### Register languages
 
@@ -57,3 +58,29 @@ or [Kramdown IAL syntax](https://kramdown.gettalong.org/syntax.html#inline-attri
 ```
 
 If you do not specify a language, then highlight.js will attempt to guess the language if `auto` is true (which it is by default).
+
+### Provide the [highlight.js] instance
+
+You can specify the `hljs` option to override the default [highlight.js] instance with your own:
+
+```js
+const hljs = require('highlight.js/lib/core')
+
+hljs.registerLanguage(
+  'javascript',
+  require('highlight.js/lib/languages/javascript')
+)
+
+const md = require('markdown-it')()
+  .use(require('markdown-it-highlightjs'), { hljs })
+```
+
+### Core plugin
+
+You may import the core `markdown-it-highlightjs` plugin directly, without any default options. You must specify an instance of [highlight.js] for the `hljs` option.
+
+```js
+const hljs = require('highlight.js/lib/core')
+const md = require('markdown-it')()
+  .use(require('markdown-it-highlightjs/core'), { hljs })
+```
