+++ package.json
@@ -14,6 +14,7 @@
   "files": [
     "README.md",
     "UNLICENSE",
+    "core.js",
     "index.js"
   ],
   "repository": {
